sagemaker/sagemaker-custom-image/requirements.txt
flask
gunicorn

# NumPy 1.x FIRST - prevents conflicts
numpy>=1.24.0,<2.0.0

# Install PyTorch with CUDA 11.8 BEFORE chandra-ocr
# This forces chandra-ocr to use our PyTorch version
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# Transformers with Qwen3VL support - required by Chandra
transformers>=4.46.0

# chandra-ocr will be installed separately in Dockerfile with --no-deps flag
# to prevent it from overriding our PyTorch CUDA 11.8 installation

# Install chandra-ocr dependencies manually (without torch/torchvision)
pillow>=9.0.0
accelerate>=0.26.0
qwen-vl-utils
huggingface-hub
